requirements.txt
acres==0.3.0
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.1.0
croniter==6.0.0
cryptography==44.0.2
dataclasses-json==0.6.7
debtcollector==3.0.0
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
docx==0.2.4
dogpile.cache==1.3.4
etelemetry==0.3.1
eval_type_backport==0.2.2
eventlet==0.39.0
faiss-cpu==1.10.0
fasteners==0.19
filelock==3.17.0
fitz==0.0.1.dev2
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
fsspec==2025.2.0
futurist==3.1.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
idna==3.10
importlib_metadata==8.6.1
iso8601==2.1.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keystoneauth1==5.10.0
keystonemiddleware==10.9.0
kombu==5.4.2
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.41
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.6
langsmith==0.3.11
logutils==0.3.5
looseversion==1.3.0
lxml==5.3.1
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
mistral-lib==3.3.1
mistralai==1.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
netaddr==1.3.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.3
openai==1.65.3
orjson==3.10.15
os-service-types==1.7.0
oslo.cache==3.10.1
oslo.concurrency==7.1.0
oslo.config==9.7.1
oslo.context==5.7.1
oslo.db==17.2.1
oslo.i18n==6.5.1
oslo.log==7.1.0
oslo.messaging==16.1.0
oslo.metrics==0.11.0
oslo.middleware==6.3.1
oslo.policy==4.5.1
oslo.serialization==5.7.0
oslo.service==4.1.1
oslo.utils==8.2.0
osprofiler==4.2.0
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
Paste==3.10.1
PasteDeploy==3.1.0
pathlib==1.0.1
pbr==6.1.1
pecan==1.5.1
pillow==11.1.0
ply==3.11
prettytable==3.15.1
prometheus_client==0.21.1
propcache==0.3.0
prov==2.0.1
psutil==7.0.0
psycopg2-binary==2.9.10
puremagic==1.28
pycadf==4.0.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydot==3.0.4
PyJWT==2.10.1
PyMuPDF==1.25.3
PyNaCl==1.5.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-keystoneclient==5.6.0
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
referencing==0.36.2
regex==2024.11.6
repoze.lru==0.7
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
Routes==2.5.1
rpds-py==0.23.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.2
simplegeneric==0.8.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
statsd==4.0.1
stevedore==5.4.1
sympy==1.13.1
tenacity==9.0.0
testresources==2.0.1
testscenarios==0.5.0
testtools==2.7.2
threadpoolctl==3.5.0
tokenizers==0.21.0
tooz==6.3.0
torch==2.6.0
tqdm==4.67.1
traits==7.0.2
transformers==4.49.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
WebOb==1.8.9
Werkzeug==3.1.3
wrapt==1.17.2
WSME==0.12.1
yappi==1.6.10
yaql==3.0.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
